Refresh-Token validation & CORS Security configuration

+++ src/main/java/org/zerock/voteservice/adapter/in/web/orchestrator/UserCreateOrchestrator.java
@@ -77,7 +77,7 @@ public class UserCreateOrchestrator extends AbstractOrchestrator<UserRegisterWeb
             return this.userRegisterService.getFailureResponseEntity(userSavedServiceResult);
         }
 
-        Integer savedUserUid = userSavedServiceResult.getUid();
+        Long savedUserUid = userSavedServiceResult.getUid();
         String savedUsername = userSavedServiceResult.getUsername();
 
         log.debug("{}User successfully saved to DB: [UID: {}]", logPrefix, savedUserUid);
